Set all grades to be over 10

+++ database/migrations/2022_01_18_183924_create_tasks_table.php
@@ -18,7 +18,7 @@ class CreateTasksTable extends Migration
             $table->string('name');
             $table->double('date_start')->nullable();
             $table->double('date_handover')->nullable();
-            $table->integer('mark')->nullable();
+            $table->float('mark')->nullable();
             $table->text('description')->nullable();
             $table->boolean('completed')->nullable();
             $table->string('google_id')->nullable();
